feat: Modified the consumer to suitably handle the chunk processing and update the transcription of the video (even if it's the first chunk being processed or not) by accepting an 'is_first_chunk' message value

+++ consumer.py
@@ -9,7 +9,7 @@ from main.models import *
 from django.conf import settings
 
 import pika, json, tempfile
-from moviepy.editor import VideoFileClip, concatenate_videoclips
+from moviepy.editor import AudioFileClip, VideoFileClip, concatenate_videoclips
 
 import requests
 
@@ -30,51 +30,61 @@ def transcribe_audio(audio_tempfile_path):
     if hasattr(response, 'json'):
         return response.json().get('text')
 
+
 def dequeue_append_video(channel, method, properties, body):
     try:
         message = json.loads(body)
         print(f'Received a message: \n{json.dumps(message, indent=4)}')
-        video_pk, chunk_file_path = message.values()
+        video_pk, chunk_file_path, is_first_chunk = message.values()
 
         video = Video.objects.get(pk=video_pk)
-        video_file_path = str(settings.BASE_DIR) + video.video_file.url
 
-        with tempfile.NamedTemporaryFile(dir=TEMPFILE_DIR, delete=False, suffix='.mp4') as final_clip_tempfile:
-            existing_clip = VideoFileClip(video_file_path)
-            new_clip = VideoFileClip(chunk_file_path)
+        audio_tempfile_path = TEMPFILE_DIR / f'audio_{video_pk}.mp3'
 
-            final_clip = concatenate_videoclips([existing_clip, new_clip])
-            final_clip_tempfile_path = final_clip_tempfile.name
-            audio_tempfile_path = TEMPFILE_DIR / f'audio_{video_pk}.mp3'
-            final_clip.write_videofile(
-                final_clip_tempfile_path,
-                temp_audiofile=audio_tempfile_path,
-                remove_temp=False,
-            )
+        if is_first_chunk:
+            video.video_file.save(f'{video_pk}.mp4', open(chunk_file_path, 'rb'))
+            audio_clip = AudioFileClip(chunk_file_path)
+            audio_clip.write_audiofile(audio_tempfile_path)
 
-            final_clip_tempfile.seek(0)
-            if os.path.exists(video_file_path):
-                os.remove(video_file_path)
-            video.video_file.save(f'{video_pk}.mp4', final_clip_tempfile)
+        else:
+            video_file_path = str(settings.BASE_DIR) + video.video_file.url
+            with tempfile.NamedTemporaryFile(dir=TEMPFILE_DIR, delete=False, suffix='.mp4') as final_clip_tempfile:
+                existing_clip = VideoFileClip(video_file_path)
+                new_clip = VideoFileClip(chunk_file_path)
 
-            final_clip_tempfile.close()
-            del existing_clip, new_clip, final_clip, final_clip_tempfile
-            if os.path.exists(final_clip_tempfile_path):
-                os.remove(final_clip_tempfile_path)
+                final_clip = concatenate_videoclips([existing_clip, new_clip])
+                final_clip_tempfile_path = final_clip_tempfile.name
 
-        # transcribe the video and update the transcription
+                final_clip.write_videofile(
+                    final_clip_tempfile_path,
+                    temp_audiofile=audio_tempfile_path,
+                    remove_temp=False,
+                )
+
+                final_clip_tempfile.seek(0)
+                if os.path.exists(video_file_path):
+                    os.remove(video_file_path)
+                video.video_file.save(f'{video_pk}.mp4', final_clip_tempfile)
+
+                final_clip_tempfile.close()
+                del existing_clip, new_clip, final_clip, final_clip_tempfile
+                if os.path.exists(final_clip_tempfile_path):
+                    os.remove(final_clip_tempfile_path)
+
+        print('Getting a transcription...')
         transcription = transcribe_audio(audio_tempfile_path)
         if not transcription:
-            print('No transcription found')
+            print('Transcription not found.')
             return
         video.transcription = transcription
         video.save()
-        print('Transcription updated successfully')
+        print('Transcription successfully updated.')
         if os.path.exists(audio_tempfile_path):
             os.remove(audio_tempfile_path)
 
     except Exception as e:
-        print(f"An error occurred during message processing: \n{'-'*44}\n\n {e}")
+        print(f"An error occurred during message processing: \n{'-'*44}\n\n {e}\n\n")
+
 
 def create_rabbitmq_connection():
     while True:
